chore(Livechat): move sdk loader to different file
